// server/src/types/repository.types.ts
export interface RepositoryInfo {
  path: string;
  name: string;
  currentBranch: string;
  isClean: boolean;
  modifiedFilesCount: number;
  untrackedFilesCount: number;
  lastCommit?: {
    hash: string;
    message: string;
    author: string;
    date: string;
  };
}

export interface BranchInfo {
  name: string;
  isCurrent: boolean;
  isRemote: boolean;
  commitHash: string;
  lastCommitDate?: string;
  lastCommitMessage?: string;
}

export interface GitStatusSummary {
  isClean: boolean;
  staged: string[];
  modified: string[];
  untracked: string[];
  conflicted: string[];
}